notebooks showing issues
In general, the whole approach seems to be going in the right direction, but I still phase certain issues that I want to adress in the future meetings. These are to be found in the new folder 'show notebooks'. Somethings weird with the harris-setting and the fourier transform integration simply doesn't work as it should and when I try to narrow it down to a 'provable problem' with as less params as possible, it starts to work. wtf?