last commit of the day
So, I found out that the integration transitions into bump solutions still look weird and still appear in parametrizations that shoudln't be turing unstable according to the current analysis. But I think, I had a mistake in the determination of fourier transformations, because I didn't consider what normalization-/osicllatory factor I had to use for the f_kernel versions. So I will determine this again and implement the turing stability analysis monday or so...