infinite spatial domain
I went from normalization of the kernel integral over a finite space to normalization over an infinite space (implementation of this is in the kernel-function, not in the ringValues() fct) Completly focus on Gaussian now. Further I applied the inverse normalisation factor of the fourier transform to the resulting f_kernel fct to ensure that it completly matches the turing analysis method using the fourier trick.