First structured analysis
see test_analysis.ipynb notebook, there are a lot of beauty apects to work on for this analysis, but generally it starts to look good. One more thing: The pattern analysis returns correctly 'spatial' patterns, when you test it individually but not in the for-loop. There might be an error there. not with the pattern collection itself