finalisation of vvfs
As long as I am in the symmetric all-to-all connectivity case, all that is implemented is done and can be tested. Next steps are to include boundary conditions and possibly heterogeneous kernels (to mimick something of short vs long maybe?! or to get assymetry?!) Another interesting factor could be the inclusion of noise & delays.